refactor: split ctf/services/range.py into a package (SonarCloud S104)
Convert the 746-line ctf/services/range.py into a ctf/services/range/
package split by responsibility: provision (single-participant provisioning
+ retry), batch (event-level throttled provisioning), tasks (spin-up task
enqueue + progress), lifecycle (stop/start/restart/destroy + cleanup), and
status (range status reads), behind a facade that re-exports the public
service surface. Extract the throttle pacing policy into a pure
compute_throttle_delay seam. No behaviour change.

Tests that reached internal range seams are rewritten to drive outcomes
through real data (DB-backed throttled-loop and force-delete tests) instead
of relocating first-party mocks, keeping the ADR-019 boundary-mock baseline
satisfied without an exception.
