+++ changelog.d/890.changed.md
@@ -1,0 +1,4 @@
+Internal refactor: `ctf/services/range.py` is now the `ctf/services/range/`
+package (`provision`, `batch`, `tasks`, `lifecycle`, `status` submodules behind
+a public facade), clearing SonarCloud `python:S104`. The throttle pacing policy
+is extracted into a pure `compute_throttle_delay` seam. No behaviour change.
